+++ exporter/exporterhelper/internal/persistent_queue.go
@@ -12,9 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build enable_unstable
-// +build enable_unstable
-
 package internal // import "go.opentelemetry.io/collector/exporter/exporterhelper/internal"
 
 import (
@@ -35,7 +32,7 @@ type persistentQueue struct {
 	stopOnce   sync.Once
 	stopChan   chan struct{}
 	numWorkers int
-	storage    persistentStorage
+	storage    *persistentContiguousStorage
 }
 
 // buildPersistentStorageName returns a name that is constructed out of queue name and signal type. This is done
@@ -55,21 +52,17 @@ func NewPersistentQueue(ctx context.Context, name string, signal config.DataType
 
 // StartConsumers starts the given number of consumers which will be consuming items
 func (pq *persistentQueue) StartConsumers(num int, callback func(item interface{})) {
-	factory := func() consumer {
-		return consumerFunc(callback)
-	}
 	pq.numWorkers = num
 
 	for i := 0; i < pq.numWorkers; i++ {
 		pq.stopWG.Add(1)
 		go func() {
 			defer pq.stopWG.Done()
-			itemConsumer := factory()
 
 			for {
 				select {
 				case req := <-pq.storage.get():
-					itemConsumer.consume(req)
+					callback(req)
 				case <-pq.stopChan:
 					return
 				}