+++ exporter/exporterhelper/internal/persistent_storage.go
@@ -12,9 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//go:build enable_unstable
-// +build enable_unstable
-
 package internal // import "go.opentelemetry.io/collector/exporter/exporterhelper/internal"
 
 import (
@@ -28,18 +25,6 @@ import (
 
 	"go.opentelemetry.io/collector/extension/experimental/storage"
 )
-
-// persistentStorage provides an interface for request storage operations
-type persistentStorage interface {
-	// put appends the request to the storage
-	put(req PersistentRequest) error
-	// get returns the next available request; note that the channel is unbuffered
-	get() <-chan PersistentRequest
-	// size returns the current size of the persistent storage with items waiting for processing
-	size() uint64
-	// stop gracefully stops the storage
-	stop()
-}
 
 // persistentContiguousStorage provides a persistent queue implementation backed by file storage extension
 //